refactor: magic number 제거
