+++ apps/web/src/app/[locale]/login/action/login.ts
@@ -4,9 +4,11 @@ import { AuthError } from 'next-auth'
 
 import { signIn } from '@/lib/nextAuth'
 
+import { LOGIN_METHOD } from '../variable'
+
 // eslint-disable-next-line consistent-return
 export async function authenticate(prevState: any, formData: FormData) {
-  const signupMethod = formData.get('signup_method')
+  const signupMethod = formData.get(LOGIN_METHOD)
 
   try {
     await signIn(signupMethod, formData)
