+++ apps/web/src/app/[locale]/login/variable.ts
@@ -1,0 +1,1 @@
+export const LOGIN_METHOD = 'SIGNUP_METHOD'
